requirements.txt
alabaster==0.7.9
alembic==0.8.9
aniso8601==1.2.0
Babel==2.3.4
bleach==1.5.0
click==6.6
decorator==4.0.10
docutils==0.13.1
entrypoints==0.2.2
Flask==0.11.1
Flask-HTTPAuth==3.2.1
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
ForgeryPy3==0.3.1
future==0.16.0
GeoAlchemy2==0.4.0
google-auth==0.5.0
google-auth-httplib2==0.0.2
google-cloud-core==0.22.1
google-cloud-storage==0.22.0
googleapis-common-protos==1.5.0
gunicorn==19.6.0
html5lib==0.9999999
httplib2==0.9.2
imagesize==0.7.1
ipykernel==4.5.2
ipyparallel==5.2.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-core==4.2.1
Mako==1.0.6
MarkupSafe==0.23
marshmallow==2.10.4
mistune==0.7.3
nbconvert==5.0.0
nbformat==4.2.0
nose==1.3.7
notebook==4.3.0
numpy==1.11.2
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
protobuf==3.1.0.post1
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycurl==7.43.0
Pygments==2.1.3
pyresttest==1.7.1
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.12.4
rsa==3.4.2
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.1
SQLAlchemy==1.1.4
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
trueskill==0.4.4
wcwidth==0.1.7
webargs==1.5.1
Werkzeug==0.11.11
widgetsnbextension==1.2.6
